# docs/swagger.yaml
definitions:
  models.Movie:
    properties:
      description:
        type: string
      director:
        type: string
      genre:
        type: string
      id:
        type: integer
      release_year:
        type: integer
      title:
        type: string
    type: object
  security.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  security.RegisterInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  security.UpdateMeInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
  /movies:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Liste tous les films
      tags:
      - movies
    post:
      consumes:
      - application/json
      parameters:
      - description: Film à créer
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/models.Movie'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Crée un film
      tags:
      - movies
  /movies/{id}:
    delete:
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Supprime un film
      tags:
      - movies
    get:
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Movie'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Récupère un film par ID
      tags:
      - movies
    put:
      consumes:
      - application/json
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      - description: Film mis à jour
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/models.Movie'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Met à jour un film
      tags:
      - movies
  /search:
    get:
      parameters:
      - description: Terme de recherche
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Recherche des films sur tous les champs
      tags:
      - movies
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email et mot de passe
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/security.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Connexion utilisateur
      tags:
      - auth
  /user/me:
    put:
      consumes:
      - application/json
      parameters:
      - description: Nouvelles données
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/security.UpdateMeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Met à jour son propre profil
      tags:
      - auth
  /user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Données d'inscription
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/security.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Inscription d'un nouvel utilisateur
      tags:
      - auth
swagger: "2.0"